Changed to right directory of test.
